Adding ExtendedClientConfiguration parameter 'retainS3Messages' that when set to 'true' prevents the companion S3 object of a message being removed from S3 on the sqs 'deleteMessage()' action. The S3 bucket will then retain a log of all messages (if the 'alwaysThroughS3' is also set to true) after they have been processed and deleted.
Also added the S3KeyGenerator interface and a default implementation to the ExtendedClientConfiguration class. The S3KeyGenerator interface allows custom implementations to add structure to the S3 bucket objects create (such as partionting on date created).

+++ src/main/java/com/amazon/sqs/javamessaging/AmazonSQSExtendedClient.java
@@ -540,7 +540,8 @@ public class AmazonSQSExtendedClient extends AmazonSQSExtendedClientBase impleme
 		String receiptHandle = deleteMessageRequest.getReceiptHandle();
 		String origReceiptHandle = receiptHandle;
 		if (isS3ReceiptHandle(receiptHandle)) {
-			deleteMessagePayloadFromS3(receiptHandle);
+			if (!clientConfiguration.isRetainS3Messages())
+				deleteMessagePayloadFromS3(receiptHandle);
 			origReceiptHandle = getOrigReceiptHandle(receiptHandle);
 		}
 		deleteMessageRequest.setReceiptHandle(origReceiptHandle);
@@ -817,7 +818,8 @@ public class AmazonSQSExtendedClient extends AmazonSQSExtendedClientBase impleme
 			String receiptHandle = entry.getReceiptHandle();
 			String origReceiptHandle = receiptHandle;
 			if (isS3ReceiptHandle(receiptHandle)) {
-				deleteMessagePayloadFromS3(receiptHandle);
+				if (!clientConfiguration.isRetainS3Messages())
+					deleteMessagePayloadFromS3(receiptHandle);
 				origReceiptHandle = getOrigReceiptHandle(receiptHandle);
 			}
 			entry.setReceiptHandle(origReceiptHandle);
@@ -1084,7 +1086,7 @@ public class AmazonSQSExtendedClient extends AmazonSQSExtendedClientBase impleme
 
 		checkMessageAttributes(batchEntry.getMessageAttributes());
 
-		String s3Key = UUID.randomUUID().toString();
+		String s3Key = clientConfiguration.getS3KeyGenerator().generateObjectKey(batchEntry);
 
 		// Read the content of the message from message body
 		String messageContentStr = batchEntry.getMessageBody();
@@ -1117,7 +1119,7 @@ public class AmazonSQSExtendedClient extends AmazonSQSExtendedClientBase impleme
 
 		checkMessageAttributes(sendMessageRequest.getMessageAttributes());
 
-		String s3Key = UUID.randomUUID().toString();
+		String s3Key = clientConfiguration.getS3KeyGenerator().generateObjectKey(sendMessageRequest);
 
 		// Read the content of the message from message body
 		String messageContentStr = sendMessageRequest.getMessageBody();
